수정 : 클래스 삭제 버그 수정 및 CloudFileHandler 예외처리 추가

+++ src/main/java/com/flo/alwaysbom/util/CloudFileHandler.java
@@ -57,8 +57,12 @@ public class CloudFileHandler implements FileHandler {
     }
 
     public boolean deleteFile(String path) {
-        String fileName = path.substring(path.lastIndexOf("/")+1, path.lastIndexOf("?")).replaceAll("%2F", "/");
-        Blob blob = storage.get(BUCKET_NAME, fileName);
-        return blob.delete();
+        if (path != null) {
+            String fileName = path.substring(path.lastIndexOf("/")+1, path.lastIndexOf("?")).replaceAll("%2F", "/");
+            Blob blob = storage.get(BUCKET_NAME, fileName);
+            return blob.delete();
+        } else {
+            return false;
+        }
     }
 }
